Исправления в сущностях и их конфигурациях в связя с изменением связей в БД. + Немного поиграл с UI (сделал круглые края у кнопки :D)

+++ Stroyzaschita.Domain/Entities/Document.cs
@@ -3,7 +3,7 @@
 public class Document {
     public short Id { get; set; }
     public byte[] File { get; set; } = default!;
-    public int? DocumentCategoryId { get; set; }
+    public short? DocumentCategoryId { get; set; }
     public DateTime AddedAt { get; set; }
 
     public DocumentCategory? DocumentCategory { get; set; }
